drawgraph: changed A, B, C, D to double

+++ drawgraph.c
@@ -3,7 +3,7 @@
 #include <string.h>
 #include <math.h>
 
-void drawGraph_body(unsigned char *pPixelBuffer, int width, int height, int A, int B, int C, int D, double S){
+void drawGraph_body(unsigned char *pPixelBuffer, int width, int height, double A, double B, double C, double D, double S){
 	int *bufi = (int *)pPixelBuffer;
 
 	// Draw x line
