+++ drawgraph.h
@@ -1,6 +1,6 @@
 #ifndef __DRAWGRAPH_H
 #define __DRAWGRAPH_H
 
-void drawGraph_body(unsigned char *pPixelBuffer, int width, int height, int A, int B, int C, int D, double S);
+void drawGraph_body(unsigned char *pPixelBuffer, int width, int height, double A, double B, double C, double D, double S);
 
 #endif //__DRAWGRAPH_H
